+++ pcap4j-packetfactory-static/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module org.pcap4j.packetfactory.statik {
+  requires org.pcap4j.core;
+  requires java.sql;
+
+  provides org.pcap4j.packet.factory.PacketFactoryBinderProvider with
+      org.pcap4j.packet.factory.statik.services.StaticPacketFactoryBinderProvider;
+}