+++ pcap4j-sample/src/main/java/module-info.java
@@ -1,0 +1,3 @@
+module org.pcap4j.sample {
+  requires org.pcap4j.core;
+}